add sample-config.yaml
