+++ script/config/sample-config.yaml
@@ -1,0 +1,15 @@
+bind: 127.0.0.1:50000
+endpoint_url: http://127.0.0.1:50000
+ecdsa_key: your_ecdsa_key
+ice_servers: stun://stun.l.google.com:19302
+stabilize_timeout: 20
+external_ip: null
+backend:
+- name: ipfs
+  prefix: http://127.0.0.1:8000
+data_storage:
+  path: /home/user/.rings/data
+  capacity: 200000000
+measure_storage:
+  path: /home/user/.rings/measure
+  capacity: 200000000
